abstract parts of ExecutionEngine and TcExecutor
