+++ include/tc/core/cuda/cuda_tc_executor.h
@@ -18,6 +18,7 @@
 #include "tc/core/halide_utils.h"
 #include "tc/core/mapping_options.h"
 #include "tc/core/polyhedral/scop.h"
+#include "tc/core/tc_executor.h"
 #include "tc/core/utils/dlpack.h"
 #include "tc/lang/parser.h"
 
@@ -25,27 +26,22 @@
 
 namespace tc {
 
-class CudaTcExecutor {
+class CudaTcExecutor : public ::tc::TcExecutor {
  public:
   CudaTcExecutor(
-      const std::string& TCDefinition,
-      const std::vector<const DLTensor*>& inputsInfo);
+      const std::string& def,
+      const std::vector<const DLTensor*>& inputsInfo)
+      : TcExecutor(def, inputsInfo) {}
   CudaTcExecutor(
-      lang::TreeRef TCDefinition,
-      const std::vector<const DLTensor*>& inputsInfo);
-  ~CudaTcExecutor();
+      lang::TreeRef tree,
+      const std::vector<const DLTensor*>& inputsInfo)
+      : TcExecutor(tree, inputsInfo) {}
+  ~CudaTcExecutor() {}
 
   CudaTcExecutor(CudaTcExecutor&&) = delete;
   CudaTcExecutor& operator=(CudaTcExecutor&&) = delete;
   CudaTcExecutor(const CudaTcExecutor&) = delete;
   CudaTcExecutor& operator=(const CudaTcExecutor&) = delete;
-
-  // Given a Tc and a list of input tensors that match the definition in the
-  // Tc in positional order, this generates the output tensor infos issued
-  // from forward inference.
-  // The typical flow is to infer output sizes, allocate/resize them within
-  // you favorite ML framework / tensor library and then call compile.
-  std::vector<const DLTensor*> inferOutputTensorInfo();
 
   // Can only be called once with specific kernel options.  Input sizes are
   // set up as constructor argument and output sizes are inferred.
@@ -54,7 +50,12 @@ class CudaTcExecutor {
   // options then just instantiate another CudaTcExecutor.
   // This is because for the time being we fully specialize all the sizes and
   // strides at runtime.
+  // @{
+  void compile(const std::string& options) override {
+    compile(MappingOptions(options));
+  }
   void compile(const tc::MappingOptions& options);
+  // @}
 
   // Run can be called multiple times given a compilation, inputs are allowed
   // to change in that their data pointer is allowed to change.
@@ -102,19 +103,6 @@ class CudaTcExecutor {
  private:
   void compileWithTcMapper();
 
-  struct TcExecutionInfo {
-    std::string kernelName;
-    std::vector<dlutils::DLTensorUPtr> inputsInfo;
-    std::vector<dlutils::DLTensorUPtr> outputsInfo;
-    std::vector<int> kernelParams;
-    std::string kernelSpecializedName;
-    std::unique_ptr<tc::MappingOptions> options;
-    std::string cudaSource;
-    Grid grid{{0, 0, 0}};
-    Block block{{0, 0, 0}};
-    std::shared_ptr<CudaRTCFunction> rtcFun;
-  };
-
  public:
   Grid grid() const {
     return execInfo_.grid;
@@ -122,16 +110,6 @@ class CudaTcExecutor {
   Block block() const {
     return execInfo_.block;
   }
-
-  const static size_t InvalidHandle = std::numeric_limits<size_t>::max();
-
- private:
-  void checkInputsCompliant(
-      const std::vector<const DLTensor*>& inputsInfo) const;
-  tc2halide::HalideComponents halideComponents_;
-  TcExecutionInfo execInfo_;
-  lang::TreeRef tcTree_;
-  mutable isl::ctx ctx_;
 };
 
 } // namespace tc
